moved widgets task to tasks folder
