+++ Gruntfile.js
@@ -1,5 +1,3 @@
-var _ = require('lodash');
-
 /* globals module,process */
 module.exports = function(grunt) {
 	var path = require("path");
@@ -207,37 +205,6 @@ module.exports = function(grunt) {
 	grunt.loadNpmTasks("grunt-contrib-compress");
 
 	grunt.loadTasks("tasks");
-
-	grunt.registerTask("widgets", "Displays a summary of widgets.", function (arg) {
-		// read all the manifests
-		var widgets = _.map(grunt.file.expand("./app/widgets/**/manifest.json"), function (path) {
-			var json = grunt.file.readJSON(path);
-			return {
-				id: json.id,
-				name: json.name,
-				legacy: false,
-				path: path
-			};
-		});
-		// read all the legacy widgets
-		var legacy = _.map(grunt.file.expand("./app/widgets/*.js"), function (path) {
-			var js = grunt.file.read(path);
-			var id = js.match(/\sid:\s(\d+),/);
-			var name = js.match(/\snicename:\s["'](\w+)["'],/);
-			return {
-				id: id ? parseInt(id[1]) : 0,
-				name: name ?  name[1] : "n/a",
-				legacy: true,
-				path: path
-			};
-		});
-		// print a report
-		_.each(_.sortBy(_.flatten([widgets, legacy]), "id"), function (manifest) {
-			grunt.log.writeln(" id:" + manifest.id + " name:" + manifest.name + " legacy:" + manifest.legacy);
-			grunt.log.verbose.writeln(manifest.path);
-		});
-	});
-
 
 	grunt.registerTask("default", [
 		"jshint:all",
